+++ Panels/FlexiblePopupPanels.js
@@ -7,7 +7,7 @@ import PopupDecorator from './PopupDecorator';
 import Viewport from './Viewport';
 
 
-const MiniPanelsDecorator = compose(
+const FlexiblePopupPanelsDecorator = compose(
 	Skinnable,
 	PopupDecorator({
 		className: 'panels mini',
@@ -22,16 +22,16 @@ const MiniPanelsDecorator = compose(
  * with the Header positioned outside the Panel background area. This is typically used for a single
  * setting or control at a time, for maximum background area viewing.
  *
- * @class MiniPanels
+ * @class FlexiblePopupPanels
  * @memberof sandstone/Panels
  * @ui
  * @public
  */
-const MiniPanels = MiniPanelsDecorator(Viewport);
+const FlexiblePopupPanels = FlexiblePopupPanelsDecorator(Viewport);
 
 // Directly set the defaultProps for position to the left side so it initially draws on the correct
 // side. The real default is assigned in PopupDecorator, but should still be overridable by an app.
-MiniPanels.defaultProps = {position: 'left'};
+FlexiblePopupPanels.defaultProps = {position: 'left'};
 
-export default MiniPanels;
-export {MiniPanels};
+export default FlexiblePopupPanels;
+export {FlexiblePopupPanels};