+++ Panels/index.js
@@ -19,7 +19,7 @@ import ActivityPanels from './ActivityPanels';
 import AlwaysViewingPanels from './AlwaysViewingPanels';
 import Breadcrumb from './Breadcrumb';
 import Header, {HeaderBase} from './Header';
-import MiniPanels from './MiniPanels';
+import FlexiblePopupPanels from './FlexiblePopupPanels';
 import OptionPanels from './OptionPanels';
 import Panel from './Panel';
 import Panels from './Panels';
@@ -32,7 +32,7 @@ export {
 	Breadcrumb,
 	Header,
 	HeaderBase,
-	MiniPanels,
+	FlexiblePopupPanels,
 	OptionPanels,
 	Panel,
 	Panels,