+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The following is a curated list of changes in the Enact sandstone module, newest
 
 ### Added
 
-- `sandstone/Panels.MiniPanels` for a flexible size pop-up Panels experience
+- `sandstone/Panels.FlexiblePopupPanels` for a flexible size pop-up Panels experience
 
 ### Fixed